+++ dotnet/tools/builder/Builder.csproj
@@ -19,6 +19,9 @@
     <OutputType>Exe</OutputType>
     <LangVersion>9.0</LangVersion>
     <NoWarn>$(NoWarn);NU1603</NoWarn>
+    <Nullable>enable</Nullable>
+    <RootNamespace>RulesMSBuild.Tools.Builder</RootNamespace>
+    <NoWarn>$(NoWarn);CS8618</NoWarn>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Build.Locator" Version="1.4.1" />
