+++ dotnet/tools/builder/Launcher/LaunchDataWriter.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 
-namespace RulesMSBuild.Tools.Builder
+namespace RulesMSBuild.Tools.Builder.Launcher
 {
     public class LaunchDataWriter
     {
