+++ dotnet/tools/builder/Launcher/LauncherFactory.cs
@@ -1,7 +1,7 @@
 using System;
 using System.IO;
 
-namespace RulesMSBuild.Tools.Builder
+namespace RulesMSBuild.Tools.Builder.Launcher
 {
     /// <summary>
     /// https://docs.google.com/document/d/1z6Xv95CJYNYNYylcRklA6xBeesNLc54dqXfri0z0e14/edit#heading=h.ehp217t4xp3w
